Av 28 possible bug block length max has rows max block columns (#115)
* Strong Typing in `Row` & `Col`
* Update API

+++ rpc/kate-rpc/src/lib.rs
@@ -160,9 +160,9 @@ where
 				.map_err(|e| internal_err!("Babe VRF not found for block {}: {:?}", block_id, e))?;
 
 			let (_, block, block_dims) = kate::com::flatten_and_pad_block(
-				block_length.rows as usize,
-				block_length.cols as usize,
-				block_length.chunk_size() as usize,
+				block_length.rows,
+				block_length.cols,
+				block_length.chunk_size(),
 				&xts_by_id,
 				seed,
 			)
